Refactor page background into layout
