+++ frontend/src/components/layout/PageBackground.tsx
@@ -1,23 +1,45 @@
 import React from "react";
 
+interface PageBackgroundProps {
+  /**
+   * Whether to show the subtle grid pattern.
+   * Default: true
+   */
+  withGrid?: boolean;
+  /**
+   * Whether to show the subtle noise texture.
+   * Default: true
+   */
+  withNoise?: boolean;
+}
+
 /**
- * PageBackground renders the static background overlays for all pages.
- * Usage example:
- *   <PageBackground />
+ * PageBackground renders the static background overlays for pages.
+ * Minimal: solid black background with an optional subtle architectural grid.
  */
-const PageBackground: React.FC = () => (
+const PageBackground: React.FC<PageBackgroundProps> = ({
+  withGrid = true,
+  withNoise = true,
+}) => (
   <>
     {/* Base background */}
-    <div className="absolute inset-0 bg-background" />
+    <div className="pointer-events-none fixed inset-0 z-[-1] bg-background" />
 
-    {/* Soft radial spotlight */}
-    <div className="absolute inset-0 bg-[radial-gradient(circle_at_center,var(--color-surface)_0%,transparent_70%)] opacity-60" />
+    {/* Subtle architectural grid (SaaS / Vercel style) */}
+    {withGrid && (
+      <div
+        className="pointer-events-none fixed inset-0 z-[-1] opacity-[0.02]"
+        style={{
+          backgroundImage: `linear-gradient(to right, #ffffff 1px, transparent 1px), linear-gradient(to bottom, #ffffff 1px, transparent 1px)`,
+          backgroundSize: "32px 32px",
+        }}
+      />
+    )}
 
-    {/* Subtle grid lines */}
-    <div className="absolute inset-0 bg-[linear-gradient(to_right,rgba(255,255,255,0.04)_1px,transparent_1px),linear-gradient(to_bottom,rgba(255,255,255,0.04)_1px,transparent_1px)] bg-[size:40px_40px] opacity-20" />
-
-    {/* Noise texture */}
-    <div className="pointer-events-none absolute inset-0 bg-[url('/noise.png')] opacity-30 mix-blend-overlay" />
+    {/* Noise texture (Logged-in views / Dashboard) */}
+    {withNoise && (
+      <div className="pointer-events-none fixed inset-0 z-[-1] bg-[url('/noise.png')] opacity-30 mix-blend-overlay" />
+    )}
   </>
 );
